==> PropaguleAnalysis/3_ShapeQuantificationNB.ipynb <==
# Quantification of shape during growth of gemmae

## Documentation and Imports


Created on 01-04-2021

Author: Valentin Laplaud

This code is performing measurements on the gemma contours computed by 'ShapeParametrisationNB'

This light version has only a few quantification used regularly, the complete version is an archive of all tested quantifications.

## Clean up before script start 

for element in dir():
    if element[0:2] != "__":
        del globals()[element]

import gc
gc.collect()


print('\033[1m' + '\033[4m' + '\nRunning ''ShapeQuantificationNB_Light'' \n' + '\033[0m')

# plotting stuff
import matplotlib as mpl
mpl.use('TkAgg')
%matplotlib inline

COLOR = 'white'
COLOR2 = 'black'

mpl.rcParams['text.color'] = COLOR
mpl.rcParams['axes.labelcolor'] = COLOR
mpl.rcParams['xtick.color'] = COLOR
mpl.rcParams['ytick.color'] = COLOR
mpl.rcParams['axes.edgecolor'] = COLOR

mpl.rcParams["figure.facecolor"] = COLOR2
mpl.rcParams["axes.facecolor"] = COLOR2
mpl.rcParams["savefig.facecolor"] = COLOR2
mpl.rcParams['axes.facecolor'] = COLOR2

import matplotlib.pyplot as plt
import matplotlib.path as mpltPath
from cycler import cycler
#Default colors
colorcycle = [plt.get_cmap('gist_rainbow')(1. * i/30) for i in range(30)]
mpl.rcParams['axes.prop_cycle'] = cycler(color=colorcycle)

# numbers handling
import numpy as np
import pandas as pd

# signal processing 
from scipy.signal import savgol_filter, correlate, correlation_lags
from scipy.interpolate import interp1d
from scipy.optimize import curve_fit, least_squares
from scipy.stats import ranksums

# images handling
from skimage import io
from skimage.filters import threshold_otsu, gaussian, laplace, sobel
from skimage.measure import label, regionprops, regionprops_table
from skimage.util import invert
from skimage.morphology import binary_opening, binary_closing, remove_small_holes,binary_erosion
from skimage.color import rgb2gray
from skimage.segmentation import active_contour, morphological_geodesic_active_contour,morphological_chan_vese, checkerboard_level_set, inverse_gaussian_gradient
import cv2 as cv

# to hide known warnings
import warnings

# General system functions
import os
import shutil
import sys

import time

# my functions
sys.path.append('../')
import VallapFunc as vf


## Define analysis functions

###  Mean contour over all gemma


####  In time

Computation only for the number of images of the shortest video.


#####  As mean (X,Y) position based on parametrisation

def computeMeanContourTime(StackList,CD,GD):
    
    # Number of image foe 24h
    nimg = 49
    
    npts = len(CD.loc[(CD.index == StackList[0]) & (CD['Img'] == 0),'Xparam'])

    # Storing variables
    alignedXs = np.empty((npts,len(StackList)))
    alignedXs[:] = np.nan
    alignedYs = np.empty((npts,len(StackList)))
    alignedYs[:] = np.nan
    
    for i in range(nimg):
        
        print('Computing mean contour for image : ' + str(i) + '/' + str(nimg).ljust(5), end='\r')
        
        # Not normalized
        
        # On each image, retrieve aligned contour for each gemma
        
        for ii,s in zip(range(len(StackList)),StackList):
            
            ImgShift = GD.loc[(GD.index == s) & (GD['Img'].values == 1), 'tdebShift'].values
            # ImgShift = 0 # No alignement
            
            if (i+ImgShift)<nimg:

                Xn = CD.loc[(CD.index == s) & (CD['Img'] == int(i+ImgShift)),'Xparam'] 
                Yn = CD.loc[(CD.index == s) & (CD['Img'] == int(i+ImgShift)),'Yparam']            

                alignedXs[:,ii] = Xn
                alignedYs[:,ii] = Yn

        # Computing mean of aligned contours        
        meanX = np.nanmean(alignedXs,1)
        meanY = np.nanmean(alignedYs,1)
        
        # Storing contour data
        data = {'Xparam':meanX,
                    'Yparam':meanY,
                    'Img':i*np.ones(len(meanX))} 
      
        newdata = pd.DataFrame(data=data,index = np.repeat('meanContTime',len(meanX)))
        
        CD = CD.append(newdata) # adding to global dataframe
        
        # Normalized
        
        # Storing variables
        alignedXNs = np.empty((npts,len(StackList)))
        alignedXNs[:] = np.nan
        alignedYNs = np.empty((npts,len(StackList)))
        alignedYNs[:] = np.nan
        
        # On each image, retrieve aligned contour for each gemma
        
        for ii,s in zip(range(len(StackList)),StackList):
           
            ImgShift = GD.loc[(GD.index == s) & (GD['Img'].values == 1), 'tdebShift'].values
            # ImgShift = 0 # No alignement
            
            if (i+ImgShift)<nimg:

                XnN = CD.loc[(CD.index == s) & (CD['Img'] == int(i+ImgShift)),'XparamNorm']  
                YnN = CD.loc[(CD.index == s) & (CD['Img'] == int(i+ImgShift)),'YparamNorm']            

                alignedXNs[:,ii] = XnN
                alignedYNs[:,ii] = YnN

        # Computing mean of aligned contours        
        meanXN = np.nanmean(alignedXNs,1)
        meanYN = np.nanmean(alignedYNs,1)
        
        # Storing contour data
        data = {'Xparam':meanXN,
                    'Yparam':meanYN,
                    'Img':i*np.ones(len(meanXN))} 
      
        newdata = pd.DataFrame(data=data,index = np.repeat('normMeanContTime',len(meanX)))
        
        CD = CD.append(newdata) # adding to global dataframe

    print('Computing mean contour for image : ' + str(nimg) + '/' + str(nimg).ljust(5), end='\n\n')
        
    return(CD)

print('Function defined.')

### Computation of growth and similarity 

Similarity : 
Comparison of one normalized contour in time (X(s),Y(s)) with either the first contour of the same gemma - selfSimilarity- or with the average contour -meanSimilarity-.

Growth : 
Computed locally as the distance between two contours, either from one time to the next, of from the origin to curent timepoint

def computeGrowthAndSimi(StackList,CD,GD):  
       
    StackList = ['meanContTime'] + StackList
    
    npts = len(CD.loc[(CD.index == StackList[0]) & (CD['Img'] == 0),'Xparam']) # number of contour points
    nppg = len(StackList) # number of propagules
    nimg =  int(1 + np.max(CD.loc[StackList[0], 'Img'])) # number of images
    
    # define storing variables
    GrowthMat = np.empty((nppg,nimg,npts,2)) # 4D matrix, for propagules, images, contours points, 
    # and full growth [0] / instant growth [1]
    
    SimMat = np.empty((nppg,nimg,npts,3)) # 4D matrix, for propagules, images, contours points, 
    # and self simi [0] / mean simi [1] / mean simi norm [2]
    SimMat[:] = np.nan
    
    for ni,n in zip(range(nppg),StackList):        
        
        ImgShift = GD.loc[(GD.index == n) & (GD['Img'].values == 1), 'tdebShift'].values
        # ImgShift = 0 # No alignement
        
        print('Processing ' + n + ' :')
            
        # t0 contours and normalized contours
        X0 = CD.loc[(CD.index == n) & (CD['Img'] == 0),'Xparam'].to_numpy()
        Y0 = CD.loc[(CD.index == n) & (CD['Img'] == 0),'Yparam'].to_numpy()
        Xn0 = CD.loc[(CD.index == n) & (CD['Img'] == 0),'XparamNorm'].to_numpy()
        Yn0 = CD.loc[(CD.index == n) & (CD['Img'] == 0),'YparamNorm'].to_numpy()
        
        for ti in range(nimg):
            
            print('Computing similarities and growth for image : ' + str(ti+1) + '/' + str(nimg).ljust(5), end='\r')
            
            # Contour and mean contour
            X = CD.loc[(CD.index == n) & (CD['Img'] == ti),'Xparam'].to_numpy()
            Y = CD.loc[(CD.index == n) & (CD['Img'] == ti),'Yparam'].to_numpy()
            meanX = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == ti), 'Xparam'].to_numpy()
            meanY = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == ti), 'Yparam'].to_numpy()

            # Normalized contour and mean contours
            Xn = CD.loc[(CD.index == n) & (CD['Img'] == ti),'XparamNorm'].to_numpy()
            Yn = CD.loc[(CD.index == n) & (CD['Img'] == ti),'YparamNorm'].to_numpy()
            meanXn = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == ti), 'Xparam'].to_numpy()
            meanYn = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == ti), 'Yparam'].to_numpy()

            if ti > 0:
                Xold = CD.loc[(CD.index == n) & (CD['Img'] == ti-1),'Xparam'].to_numpy()
                Yold = CD.loc[(CD.index == n) & (CD['Img'] == ti-1),'Yparam'].to_numpy()
            else:
                Xold = CD.loc[(CD.index == n) & (CD['Img'] == ti),'Xparam'].to_numpy()
                Yold = CD.loc[(CD.index == n) & (CD['Img'] == ti),'Yparam'].to_numpy()
                    
            
            ### Growth
            # Growth from original shape
            
            Gn0 = vf.dist(X,Y,X0,Y0)

            # Growth from previous timepoint
            
            Gn = vf.dist(X,Y,Xold,Yold)

            
            GrowthMat[ni,ti,:,0] = Gn0
            GrowthMat[ni,ti,:,1] = Gn

            
            ### similarities           

            # Computing self similarity
            Sn0 = vf.dist(Xn,Yn,Xn0,Yn0)
            SimMat[ni,ti,:,0] = Sn0
                        
            if (ti+ImgShift)<49:
                # Shifted contours for comparison to mean
                X = CD.loc[(CD.index == n) & (CD['Img'] == int(ti+ImgShift)),'Xparam'].to_numpy()
                Y = CD.loc[(CD.index == n) & (CD['Img'] == int(ti+ImgShift)),'Yparam'].to_numpy()

                Xn = CD.loc[(CD.index == n) & (CD['Img'] == int(ti+ImgShift)),'XparamNorm'].to_numpy()
                Yn = CD.loc[(CD.index == n) & (CD['Img'] == int(ti+ImgShift)),'YparamNorm'].to_numpy()

                # Computing mean similarity
                Snm = vf.dist(X,Y,meanX,meanY)
                SnmN = vf.dist(Xn,Yn,meanXn,meanYn)
        
                SimMat[ni,ti,:,1] = Snm
                SimMat[ni,ti,:,2] = SnmN
        
       
        print('Computing similarities and growth for image : ' + str(ti+1) + '/' + str(nimg).ljust(5), end='\n\n')    
            
    return(CD,GrowthMat,SimMat)  


print('Function defined.')       
            

## Plot fuctions

### Plot gemmae area in time


def plotArea(StackList, GD, P, CCs, Title, **kwargs):

    # Colors
    colorcycle = [plt.get_cmap('gist_rainbow')(1. * i/30) for i in range(30)]
    mpl.rcParams['axes.prop_cycle'] = cycler(color=colorcycle)
    
    #init and read kwargs
    Trap = False # default value
    soloGraphs = False
    
    for key, value in kwargs.items(): 
        if key == 'Trapped':
            Trap = value
        elif key == 'solo':
            soloGraphs = value
        else:
            print('Unknown key : ' + key+ '. Kwarg ignored.')
         
    # check existence of figure folder, if absent, create it
    if not os.path.exists(P + '\\Figures'):
            os.mkdir(P + '\\Figures') # create folder
    
    # plot figure
    
    ###################### plot individual graphs #######################
    
    if soloGraphs: 
    
        for s in StackList:
            

            fig1 = plt.figure(dpi = 200,facecolor='black')
            fig1.suptitle(s + ' - Size vs. time')
            plt.xlabel('Time (min)')
            plt.ylabel('Size (mm²)')
            

            xTime = GD.at[s, 'Time (min)'].to_numpy()[GD.at[s, 'GoodFocus']==True]
        
            
            PropBBOX = GD.at[s, 'Bounding box (px)']

            plt.plot(xTime,GD.at[s, 'Size (mm²)'].to_numpy()[GD.at[s, 'GoodFocus']==True],'-o',ms=3,figure=fig1,label=s)

            # if there is a trap, find out when th ppgls touches up and bottom wall
            if Trap:
                trapdata =vf.LoadImageJResults(P + '\\' + s + '_Trap.txt', ['BX','BY','Width','Height'])
                TrapBBOX = [trapdata.at[0, 'BY'], trapdata.at[0, 'BX'],trapdata.at[0, 'BY'] + trapdata.at[0, 'Height'] ,
                        trapdata.at[0, 'BX'] + trapdata.at[0, 'Width']]

                BBOXDiff = PropBBOX - TrapBBOX            
                idxTouch = np.argwhere(np.logical_and(BBOXDiff[0:,0]<0, BBOXDiff[0:,2]>0))

                plt.plot(xTime[idxTouch],GD.at[s, 'Size (mm²)'][idxTouch],'-or',lw=0.5,ms=1.5,figure=fig1,label='Touching UpDown')
                plt.legend(prop={'size': 4})

                fig1.savefig(P + '\\Figures\\' + s + '_SizeTime.png')    


    ###################### plot common graphs #######################
    
    fig2,ax2 = plt.subplots(dpi = 200,facecolor='black')
    fig2.suptitle(Title + ' - Area vs. time')
    plt.xlabel('Time (min)')
    plt.ylabel('Area (mm²)')
    
    fig3,ax3 = plt.subplots(dpi = 200,facecolor='black')
    fig3.suptitle(Title + ' - Norm Area vs. time')
    plt.xlabel('Time (min)')
    plt.ylabel('Area (normalized)')
    plt.legend(prop={'size': 4})
    
    # Computing mean area over all gemmae for ach image
    MeanA = np.empty(max(GD['Img']+1))
    MeanTime = np.empty(max(GD['Img']+1))
    StdA = np.empty(max(GD['Img']+1))
    
    for im in range(0,max(GD['Img']+1)):
                    
        MeanA[im] = GD[GD['Img'] == im]['Area'].to_numpy()[np.array([GD['GoodFocus']==True])[0][GD['Img'] == im]].mean()
        MeanTime[im] = im*30
        StdA[im] = GD[GD['Img'] == im]['Area'].to_numpy()[np.array([GD['GoodFocus']==True])[0][GD['Img'] == im]].std()
        nppg = len(GD[GD['Img'] == im]['Area'].to_numpy()[np.array([GD['GoodFocus']==True])[0][GD['Img'] == im]])
    
    for s in StackList:        
                   
        xTime = GD.at[s, 'Time (min)'].to_numpy()[GD.at[s, 'GoodFocus']==True]
        Area = GD.at[s, 'Area'].to_numpy()[GD.at[s, 'GoodFocus']==True]
        
        NormArea = Area/Area[0]
        
        ax2.plot(xTime,Area/1000000,'-o',ms=3,label=s)
        ax3.plot(xTime,NormArea,'-o',ms=3,label=s)
        
    #ax2.errorbar(MeanTime,MeanA/1000000,yerr=StdA/1000000/np.sqrt(nppg),ecolor='w',color='w', capsize=3)
    ax2.plot(MeanTime,MeanA/1000000,'-*w',ms=2,label='Mean')
    #ax3.errorbar(MeanTime,MeanA/MeanA[0],yerr=StdA/MeanA[0]/np.sqrt(nppg),ecolor='w',color='w', capsize=3)
    ax3.plot(MeanTime,MeanA/MeanA[0],'-*w',ms=2,label='Mean')
    
    plt.figure(fig2.number)
    plt.legend(prop={'size': 4})
    fig2.savefig(P + '\\Figures\\AreaTime_' + Title + '.png')
    
    plt.figure(fig3.number)
    plt.legend(prop={'size': 4})
    fig3.savefig(P + '\\Figures\\NormAreaTime_' + Title + '.png')
    
    
    ###################### plot time correlation #######################
    
    t0 = GD.loc[(GD.index == StackList[0]) & (GD['Img'] == 1),'CorrT0'].to_numpy()
    t1 = GD.loc[(GD.index == StackList[0]) & (GD['Img'] == 1),'CorrT1'].to_numpy()
    t2 = GD.loc[(GD.index == StackList[0]) & (GD['Img'] == 1),'CorrT2'].to_numpy()
    
    fig4,ax4 = plt.subplots(dpi = 200,facecolor='black')
    fig4.suptitle(Title + ' - Area growth time correlation (T = ' + str(int(t0[0])) + ', ' + str(int(t1[0])) + ', ' + str(int(t2[0])) +  ')' + 
                  '\nPCC = ' + str(CCs[0]) + ' - KT = ' + str(CCs[1]) + ' - SR = ' + str(CCs[2]))
    plt.xlabel('G01')
    plt.ylabel('G12')
    plt.legend(prop={'size': 4})
    
    ax4.scatter(GD['G01'],GD['G12'])
    
    fig4.savefig(P + '\\Figures\\GrowthTimeCorr_' + Title + '.png')
    plt.show()
    return    

print('Function defined')

### Plot evolution of gemma shape in time and mean gemma contour

def plotShapes(P, StackList,CD,GD, figname, **kwargs):
    
    Plots = False
    
    for key, value in kwargs.items(): 
        if key == 'showplots':
            Plots = value
        else:
            print('Unknown key : ' + key + '. Kwarg ignored.')
    
    # create save folders
                    
    if not os.path.exists(P + '\\Figures\\MeanContourTimeNorm_' + figname + '\\'):
            os.mkdir(P + '\\Figures\\MeanContourTimeNorm_' + figname + '\\')  
                    
    if not os.path.exists(P + '\\Figures\\MeanContourTimeEvol_' + figname + '\\'):
            os.mkdir(P + '\\Figures\\MeanContourTimeEvol_' + figname + '\\') 
                    
    if not os.path.exists(P + '\\Figures\\MeanContourTimeVariation_' + figname + '\\'):
            os.mkdir(P + '\\Figures\\MeanContourTimeVariation_' + figname + '\\') 
                    
    #if not os.path.exists(P + '\\Figures\\MeanPDFContourTimeNorm_' + figname + '\\'):
    #        os.mkdir(P + '\\Figures\\MeanPDFContourTimeNorm_' + figname + '\\')  
                    
    #if not os.path.exists(P + '\\Figures\\MeanPDFContourTimeEvol_' + figname + '\\'):
    #        os.mkdir(P + '\\Figures\\MeanPDFContourTimeEvol_' + figname + '\\') 
                    
    #if not os.path.exists(P + '\\Figures\\MeanContourAreaEvol_' + figname + '\\'):
    #        os.mkdir(P + '\\Figures\\MeanContourAreaEvol_' + figname + '\\') 
                    
    if not os.path.exists(P + '\\Figures\\ContourEvolTime\\'):
            os.mkdir(P + '\\Figures\\ContourEvolTime\\') 
                

    
    # Number of images
    nmint = 1 + int(np.min([np.max(CD.loc[s, 'Img']) for s in StackList]))
    """narea = 1 + int(np.max([np.max(CD.loc[s, 'AreaGroup']) for s in StackList]))"""
        
    # Colors
    colorcycle = [plt.get_cmap('jet')(1. * i/nmint) for i in range(nmint)]
    mpl.rcParams['axes.prop_cycle'] = cycler(color=colorcycle)


    ################ Mean Contours in time #################
    
    # initialize figure for mean contour
    fig0, ax0 = plt.subplots(facecolor='black')
    fig0.set_dpi(250)
    fig0.suptitle('MeanContour in time - ' + figname)
    plt.xlabel('(µm)')
    plt.ylabel('(µm)')
    plt.xlim(-800,800)
    plt.ylim(-800,800)
    ax0.plot(0,0,'*r')                               
    ax0.set_aspect('equal', adjustable='box')
    
    print('Plotting mean contours in time...', end='')
    
    for i in range(nmint):
  
        # Plot shape
        meanX = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == i),'Xparam']  
        meanY = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == i),'Yparam']
        
        ax0.plot(meanX,meanY,'-', lw=0.5)
        fig0.savefig(P + '\\Figures\\MeanContourTimeEvol_' + figname + '\\' + str(i)  + '.png')
        
        fig01, ax01 = plt.subplots(facecolor='black')
        fig01.set_dpi(250)
        fig01.suptitle('Variation of individuals around mean contour - ' + figname)
        plt.xlabel('(µm)')
        plt.ylabel('(µm)')
        plt.xlim(-800,800)
        plt.ylim(-800,800)
        ax01.plot(0,0,'*r')                               
        ax01.set_aspect('equal', adjustable='box')
        
        
        for s in StackList:
            X = CD.loc[(CD.index == s) & (CD['Img'] == i),'Xparam']  
            Y = CD.loc[(CD.index == s) & (CD['Img'] == i),'Yparam']
            
            ax01.plot(X,Y,'-',lw=0.4,color=[0.65, 0.65, 0.65])
            
        ax01.plot(meanX,meanY,'-r', lw=1.2)
        fig01.savefig(P + '\\Figures\\MeanContourTimeVariation_' + figname + '\\' + str(i)  + '.png')
        plt.close(fig01)
            
    # Save fig
    fig0.savefig(P + '\\Figures\\MeanContourTime_' + figname + '.png')
    
    
    print('Done and saved.')
    
    
    print('Plotting normalized mean contours in time...', end='')
    
    colorcyclejet = [plt.get_cmap('jet')(1. * i/nmint) for i in range(nmint)]
    
    fig02, ax02 = plt.subplots(facecolor='black')
    fig02.set_dpi(250)
    fig02.suptitle('Normalized meanContour in time - ' + figname)
    plt.xlabel('(norm)')
    plt.ylabel('(norm)')
    plt.xlim(-1.3,1.3)
    plt.ylim(-1.3,1.3)
    ax02.plot(0,0,'*r')                               
    ax02.set_aspect('equal', adjustable='box')    
    
    for i in range(nmint):
    
        # initialize figure for normalized mean contour
        fig1, ax1 = plt.subplots(facecolor='black')
        fig1.set_dpi(250)
        fig1.suptitle('Normalized meanContour in time - ' + figname)
        plt.xlabel('(norm)')
        plt.ylabel('(norm)')
        plt.xlim(-1.3,1.3)
        plt.ylim(-1.3,1.3)
        ax1.plot(0,0,'*r')                               
        ax1.set_aspect('equal', adjustable='box')     
  
        # Plot shape
        meanX = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == i),'Xparam']  
        meanY = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == i),'Yparam']
                
        ax02.plot(meanX,meanY,'-', lw=1.5, c = colorcyclejet[i])
        ax1.plot(meanX,meanY,'-', lw=1.5, c = colorcyclejet[i])
            
        # Save fig
        fig1.savefig(P + '\\Figures\\MeanContourTimeNorm_' + figname + '\\MeanContourTimeNorm_' + figname + '_' + str(i) + '.png')
        plt.close(fig1)
        
    # Save fig
        fig02.savefig(P + '\\Figures\\MeanContourTimeNormAll_' + figname + '.png')
        

    
    print('Done and saved.')
    
    """
    ################ Mean PDF Contours in time #################
    
    # initialize figure for mean contour
    fig0, ax0 = plt.subplots(facecolor='black')
    fig0.set_dpi(250)
    fig0.suptitle('Mean PDF Contour in time - ' + figname)
    plt.xlabel('(µm)')
    plt.ylabel('(µm)')
    plt.xlim(-800,800)
    plt.ylim(-800,800)
    ax0.plot(0,0,'*r')                               
    ax0.set_aspect('equal', adjustable='box')
    
    print('Plotting PDF mean contours in time...', end='') 
    
    for i in range(nmint):
  
        # Plot shape
        meanX = CD.loc[(CD.index == 'meanContPDF') & (CD['Img'] == i),'Xparam']  
        meanY = CD.loc[(CD.index == 'meanContPDF') & (CD['Img'] == i),'Yparam']
        
        ax0.plot(meanX,-meanY,'-', lw=0.5)
        fig0.savefig(P + '\\Figures\\MeanPDFContourTimeEvol_' + figname + '\\' + str(i)  + '.png')
            
    # Save fig
    fig0.savefig(P + '\\Figures\\MeanPDFContourTime_' + figname + '.png')
    plt.figure(fig0.number)
    #plt.close(fig0.number)
    
    print('Done and saved.')
    
    
    print('Plotting normalized mean PDF contours in time...', end='')
    
    colorcyclejet = [plt.get_cmap('jet')(1. * i/nmint) for i in range(nmint)]
    
    fig0, ax0 = plt.subplots(facecolor='black')
    fig0.set_dpi(250)
    fig0.suptitle('Normalized Mean PDF Contour in time - ' + figname)
    plt.xlabel('(norm)')
    plt.ylabel('(norm)')
    plt.xlim(-1.1,1.1)
    plt.ylim(-1.1,1.1)
    ax0.plot(0,0,'*r')                               
    ax0.set_aspect('equal', adjustable='box')    
    
    for i in range(nmint):
    
        # initialize figure for normalized mean contour
        fig1, ax1 = plt.subplots(facecolor='black')
        fig1.set_dpi(250)
        fig1.suptitle('Normalized mean PDF Contour in time - ' + figname)
        plt.xlabel('(norm)')
        plt.ylabel('(norm)')
        plt.xlim(-1.1,1.1)
        plt.ylim(-1.1,1.1)
        ax1.plot(0,0,'*r')                               
        ax1.set_aspect('equal', adjustable='box')     
  
        # Plot shape
        meanX = CD.loc[(CD.index == 'normMeanContPDF') & (CD['Img'] == i),'Xparam']  
        meanY = CD.loc[(CD.index == 'normMeanContPDF') & (CD['Img'] == i),'Yparam']
        
        ax0.plot(meanX,-meanY,'-', lw=1.5, c = colorcyclejet[i])
        ax1.plot(meanX,-meanY,'-', lw=1.5, c = colorcyclejet[i])
            
        # Save fig
        fig1.savefig(P + '\\Figures\\MeanPDFContourTimeNorm_' + figname + '\\' + str(i) + '.png')
        plt.figure(fig1.number)
        plt.close(fig1.number)
        
    # Save fig
        fig0.savefig(P + '\\Figures\\MeanPDFContourTimeNormAll_' + figname + '.png')
        plt.figure(fig0.number)
        plt.close(fig0.number)
    
    print('Done and saved.')
    
    ################ Mean Contours in area #################
    
    # initialize figure for mean contour
    fig0, ax0 = plt.subplots(facecolor='black')
    fig0.set_dpi(250)
    fig0.suptitle('MeanContour in area - ' + figname)
    plt.xlabel('(µm)')
    plt.ylabel('(µm)')
    plt.xlim(-800,800)
    plt.ylim(-800,800)
    ax0.plot(0,0,'*r')                               
    ax0.set_aspect('equal', adjustable='box')
    
    print('Plotting mean contours in area...', end='')
    
    
    colorcyclejet = [plt.get_cmap('jet')(1. * i/narea) for i in range(narea)]
    
    for i in range(narea):
  
        #Plot shape
        meanX = CD.loc[(CD.index == 'meanContArea') & (CD['AreaGroup'] == i),'Xparam']  
        meanY = CD.loc[(CD.index == 'meanContArea') & (CD['AreaGroup'] == i),'Yparam']
        
        ax0.plot(meanX,meanY,'-', lw=0.5,c=colorcyclejet[i])
        fig0.savefig(P + '\\Figures\\MeanContourAreaEvol_' + figname + '\\' + str(i)  + '.png')
            
    # Save fig
    fig0.savefig(P + '\\Figures\\MeanContourArea_' + figname + '.png')
    plt.figure(fig0.number)
    #plt.close(fig0.number)
    
    print('Done and saved.')
        
   """
    
    # plotting individual evolution of contours in time

    """ for s in StackList:

        print('Plotting individual contours for gemma : ' + s + '...', end='')
        
        if not os.path.exists(P + '\\Figures\\ContourEvolTime\\' + s):
                os.mkdir(P + '\\Figures\\ContourEvolTime\\' + s) 

        fig02, ax02 = plt.subplots(facecolor='black')
        fig02.set_dpi(250)
        fig02.suptitle('Contour in time - ' + s)
        plt.xlabel('(µm)')
        plt.ylabel('(µm)')
        plt.xlim(-800,800)
        plt.ylim(-800,800)
        ax02.plot(0,0,'*r')                               
        ax02.set_aspect('equal', adjustable='box')

        for i in range(int(np.max(CD.loc[s, 'Img'] + 1))):

            X = CD.loc[(CD.index == s) & (CD['Img'] == i),'Xparam']  
            Y = CD.loc[(CD.index == s) & (CD['Img'] == i),'Yparam']

            ax02.plot(X,Y,'-',lw=0.4)
            fig02.savefig(P + '\\Figures\\ContourEvolTime\\' + s + '\\' + str(i) + '.png') """

        #print('Done and saved.')
           
    if Plots:
        plt.show()
    else:
        plt.close(fig0.number)   
        plt.close(fig02.number)
        plt.close(fig1.number)
        
    return

print('Function defined.')
 

### Plot growth quantifications

def plotGrowth(P,StackList,CD,GrowthMat, **kwargs):
    
    Plots = False
    
    for key, value in kwargs.items(): 
        if key == 'showplots':
            Plots = value
        else:
            print('Unknown key : ' + key + '. Kwarg ignored.')
            
    ### Individual growth plot    
    
    if not os.path.exists(P + '\\Figures\\GrowthPlots\\'):
        os.mkdir(P + '\\Figures\\GrowthPlots\\') # create image folder
        
    if not os.path.exists(P + '\\Figures\\GrowthPlots\\FullGrowth\\'):
        os.mkdir(P + '\\Figures\\GrowthPlots\\FullGrowth\\') # create image folder  
        
    if not os.path.exists(P + '\\Figures\\GrowthPlots\\InstantGrowth\\'):
        os.mkdir(P + '\\Figures\\GrowthPlots\\InstantGrowth\\') # create image folder   
    
    nppg = len(StackList)
    
    for s, si in zip(StackList,range(nppg)):    
        
        print('Plot growth for gemma : ' + s + '...', end='')
        
        if not os.path.exists(P + '\\Figures\\GrowthPlots\\FullGrowth\\' + s  + '_FullGrowth\\'):
            os.mkdir(P + '\\Figures\\GrowthPlots\\FullGrowth\\' + s  + '_FullGrowth\\') # create save folder
        
        if not os.path.exists(P + '\\Figures\\GrowthPlots\\InstantGrowth\\' + s  + '_InstantGrowth\\'):
            os.mkdir(P + '\\Figures\\GrowthPlots\\InstantGrowth\\' + s  + '_InstantGrowth\\') # create save folder
       
        n = int(1 + np.max(CD.loc[s, 'Img'])) # number of images
        
        # ppg contour at t = 0
        X0 = CD.loc[(CD.index == s) & (CD['Img'] == 0),'Xparam'].to_numpy()
        Y0 = CD.loc[(CD.index == s) & (CD['Img'] == 0),'Yparam'].to_numpy()
        
        # max ppg size
        maxX = np.max(np.abs(CD.loc[(CD.index == s) & (CD['Img'] == n-1),'Xparam'].to_numpy()))*1.05
        maxY = np.max(np.abs(CD.loc[(CD.index == s) & (CD['Img'] == n-1),'Yparam'].to_numpy()))*1.05
        
        fig0,ax0 = plt.subplots(dpi=300)
        fig0.suptitle(s + ' - Growth from initial state')
        ax0.plot(X0,Y0,'--w',zorder = 0)
        ax0.set_aspect('equal', adjustable='box')
        ax0.set_xlim([-maxX, maxX])
        ax0.set_ylim([-maxY, maxY])
        
        
        
        for i in range(n):      
            
            # ppg contour
            X = CD.loc[(CD.index == s) & (CD['Img'] == i),'Xparam'].to_numpy()
            Y = CD.loc[(CD.index == s) & (CD['Img'] == i),'Yparam'].to_numpy()
            
            # Parametrized growth
            Gn0 = GrowthMat[si,i,:,0] # compared to initial state
            Gn = GrowthMat[si,i,:,1] # compared to previous time point
            
            
            fig1,ax1 = plt.subplots(dpi=300)
            fig1.suptitle(s + ' - Growth from initial state - Img ' + str(i))
            ax1.plot(X0,Y0,'--w',zorder = 0)
            sc1 = ax1.scatter(X,Y,c=Gn0, cmap = 'jet', s = 2)
            fig1.colorbar(sc1, ax = ax1, label = 'Growth (µm)',shrink = 0.6)
            ax1.set_aspect('equal', adjustable='box')
            ax1.set_xlim([-maxX, maxX])
            ax1.set_ylim([-maxY, maxY])
            
            fig1.tight_layout()
            fig1.savefig(P + '\\Figures\\GrowthPlots\\FullGrowth\\' + s  + '_FullGrowth\\' + str(i))
            
            if not Plots:
                plt.close(fig1)

            fig2,ax2 = plt.subplots(dpi=300)
            fig2.suptitle(s + ' - Instantaneous growth - Img ' + str(i))
            sc2 = ax2.scatter(X,Y,c=Gn, cmap = 'jet', s = 2)
            fig2.colorbar(sc2, ax = ax2, label = 'Instantaneous growth (µm)',shrink = 0.6)
            ax2.set_aspect('equal', adjustable='box')
            ax2.set_xlim([-maxX, maxX])
            ax2.set_ylim([-maxY, maxY])
            
            fig2.tight_layout()
            fig2.savefig(P + '\\Figures\\GrowthPlots\\InstantGrowth\\' + s  + '_InstantGrowth\\' + str(i))
            
            if not Plots:
                plt.close(fig2)
            
            if (i>10) & (np.mod(i,4) == 3):
                Gn0N = np.divide(Gn0,np.nanmedian(Gn0)) # normalized by median
                sc0 = ax0.scatter(X,Y,c=Gn0N, cmap = 'jet', s = 0.25)
            
        fig0.colorbar(sc0, ax = ax0, label = 'Normalized growth (median)',shrink = 0.6)
        fig0.tight_layout()
        fig0.savefig(P + '\\Figures\\GrowthPlots\\FullGrowth\\' + s  + '_FullGrowth')
            
        if not Plots:
            plt.close(fig0)
                
        if Plots:
            plt.show()
        
        print('Done and saved.\n')
    
    
    return

print('Function defined.')
    

### Plot similarities

def plotSimi(P,StackList,CD,SimMat,Title, **kwargs):
    
    Plots = False
    
    for key, value in kwargs.items(): 
        if key == 'showplots':
            Plots = value
        else:
            print('Unknown key : ' + key + '. Kwarg ignored.')
            
    ### Individual growth plot    
    
    if not os.path.exists(P + '\\Figures\\SimiPlots\\'):
        os.mkdir(P + '\\Figures\\SimiPlots\\') # create image folder
        
    if not os.path.exists(P + '\\Figures\\SimiPlots\\SelfSimi\\'):
        os.mkdir(P + '\\Figures\\SimiPlots\\SelfSimi\\') # create image folder   
        
    if not os.path.exists(P + '\\Figures\\SimiPlots\\MeanSimi\\'):
        os.mkdir(P + '\\Figures\\SimiPlots\\MeanSimi\\') # create image folder   
        
    if not os.path.exists(P + '\\Figures\\SimiPlots\\MeanSimiNorm\\'):
        os.mkdir(P + '\\Figures\\SimiPlots\\MeanSimiNorm\\') # create image folder   
    
    nppg = len(StackList)
    
    for s, si in zip(StackList,range(nppg)):    
        
        print('Plot similarities for gemma : ' + s + '...', end='')
        
        if not os.path.exists(P + '\\Figures\\SimiPlots\\SelfSimi\\' + s  + '_SelfSimi\\'):
            os.mkdir(P + '\\Figures\\SimiPlots\\SelfSimi\\' + s  + '_SelfSimi\\') # create image folder   
        
        if not os.path.exists(P + '\\Figures\\SimiPlots\\MeanSimi\\' + s  + '_MeanSimi\\'):
            os.mkdir(P + '\\Figures\\SimiPlots\\MeanSimi\\' + s  + '_MeanSimi\\') # create image folder   

        if not os.path.exists(P + '\\Figures\\SimiPlots\\MeanSimiNorm\\' + s  + '_MeanSimiNorm\\'):
            os.mkdir(P + '\\Figures\\SimiPlots\\MeanSimiNorm\\' + s  + '_MeanSimiNorm\\') # create image folder   
        
        ImgShift = GD.loc[(GD.index == s) & GD['Img'] == 1, 'tdebShift']
        # ImgShift = 0 # No alignement
       
        n = int(1 + np.max(CD.loc[s, 'Img'])) # number of images
        
        for i in range(n):      
            
            # ppg and mean contour
            X = CD.loc[(CD.index == s) & (CD['Img'] == i),'Xparam'].to_numpy()
            Y = CD.loc[(CD.index == s) & (CD['Img'] == i),'Yparam'].to_numpy()
            
            Xm = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == i),'Xparam'].to_numpy()
            Ym = CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == i),'Yparam'].to_numpy()
            
            XN = CD.loc[(CD.index == s) & (CD['Img'] == i),'XparamNorm'].to_numpy()
            YN = CD.loc[(CD.index == s) & (CD['Img'] == i),'YparamNorm'].to_numpy()
            
            XmN = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == i),'Xparam'].to_numpy()
            YmN = CD.loc[(CD.index == 'normMeanContTime') & (CD['Img'] == i),'Yparam'].to_numpy()
            
            # max ppg size
            maxX = np.max(np.abs(CD.loc[(CD.index == s) & (CD['Img'] == n-1),'Xparam'].to_numpy()))*1.1
            maxY = np.max(np.abs(CD.loc[(CD.index == s) & (CD['Img'] == n-1),'Yparam'].to_numpy()))*1.1
            
            maxXm = np.max(np.abs(CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == n-1),'Xparam'].to_numpy()))*1.1
            maxYm = np.max(np.abs(CD.loc[(CD.index == 'meanContTime') & (CD['Img'] == n-1),'Yparam'].to_numpy()))*1.1
            
            Xmax = np.max([maxX, maxXm])
            Ymax = np.max([maxY, maxYm])
            
            
            # Similarities
            Sn0 = SimMat[si+1,i,:,0]
            Snm = SimMat[si+1,i,:,1]
            SnmN = SimMat[si+1,i,:,2]
            
            fig0,ax0 = plt.subplots(dpi=300)
            fig0.suptitle(s + ' - Mean similarity - ' + Title)
            ax0.plot(Xm,Ym,'--w',zorder = 0)
            ax0.set_aspect('equal', adjustable='box')
            ax0.set_xlim([-Xmax, Xmax])
            ax0.set_ylim([-Ymax, Ymax]) 
            sc0 = ax0.scatter(X,Y,c=Snm, cmap = 'jet', s = 0.25)
            fig0.colorbar(sc0, ax = ax0, label = 'Distance to mean',shrink = 0.6)
            fig0.tight_layout()
            fig0.savefig(P + '\\Figures\\SimiPlots\\MeanSimi\\' + s  + '_MeanSimi\\' + str(i))
            
            
            fig1,ax1 = plt.subplots(dpi=300)
            fig1.suptitle(s + ' - Normalized mean similarity - ' + Title)
            ax1.plot(XmN,YmN,'--w',zorder = 0)
            ax1.set_aspect('equal', adjustable='box')
            ax1.set_xlim([-1.5, 1.5])
            ax1.set_ylim([-1.5, 1.5]) 
            sc1 = ax1.scatter(XN,YN,c=SnmN, cmap = 'jet', s = 0.25)
            fig1.colorbar(sc1, ax = ax1, label = 'Distance to normalized mean',shrink = 0.6)
            fig1.tight_layout()
            fig1.savefig(P + '\\Figures\\SimiPlots\\MeanSimiNorm\\' + s  + '_MeanSimiNorm\\' + str(i))
            
            if not Plots:
                plt.close(fig0)
                plt.close(fig1)

        print('Done and saved.\n')
            
    # box plot of distance to mean for all ppg at 0, 12 and 24h
    Snm0h  = np.median(SimMat[1:,0,:,1],axis = 1)
    Snm12h = np.median(SimMat[1:,24,:,1],axis = 1)
    Snm24h = np.median(SimMat[1:,48,:,1],axis = 1)

    fig2,ax2 = plt.subplots(dpi=300)
    fig2.suptitle('Distances to mean - ' + Title)
    plotprops = {'color':'w'}           
    flierprops = dict(marker='d', markerfacecolor='r', markersize=8,
              markeredgecolor='none')
    plt.boxplot([Snm0h, Snm12h, Snm24h], labels = ['0h', '12h', '24h'],boxprops = plotprops, capprops =plotprops,
                flierprops=flierprops,
                whiskerprops=plotprops)
    
    # stats    
    s01,p01 = ranksums(Snm0h,Snm12h)
    s12,p12 = ranksums(Snm12h,Snm24h)
    s02,p02 = ranksums(Snm0h,Snm24h)

    h01= np.max([np.max(Snm0h),np.max(Snm12h)])
    h12= np.max([np.max(Snm24h),np.max(Snm12h)])
    h02= np.max([np.max(Snm0h),np.max(Snm24h)])
    
    hmax = np.max([h01,h02,h12])
    step = hmax*0.15

    ax2.plot([1, 2], [hmax+step ,hmax+step], 'w-',zorder=0)
    ax2.text(1.5,hmax+step*1.3,'p = ' + str(round(p01*1000)/1000), ha='center',fontsize='small')
    
    ax2.plot([2, 3], [hmax+step*2 ,hmax+step*2], 'w-',zorder=0)
    ax2.text(2.5,hmax+step*2.3,'p = ' + str(round(p12*1000)/1000), ha='center',fontsize='small')
    
    ax2.plot([1, 3], [hmax+step*3 ,hmax+step*3], 'w-',zorder=0)
    ax2.text(2,hmax+step*3.3,'p = ' + str(round(p02*1000)/1000), ha='center',fontsize='small')
    
    ax2.set_ylim(top=hmax+step*4)
    
    fig2.savefig(P + '\\Figures\\SimiPlots\\Boxplot_DistancesToMean' + Title)

    # same for normalized contours
    SnmN0h  = np.median(SimMat[1:,0,:,2],axis = 1)
    SnmN12h = np.median(SimMat[1:,24,:,2],axis = 1)
    SnmN24h = np.median(SimMat[1:,48,:,2],axis = 1)

    fig3,ax3 = plt.subplots(dpi=300)
    fig3.suptitle('Distances to normalized mean - ' + Title)
    plotprops = {'color':'w'}           
    flierprops = dict(marker='d', markerfacecolor='r', markersize=8,
              markeredgecolor='none')
    plt.boxplot([SnmN0h, SnmN12h, SnmN24h], labels = ['0h', '12h', '24h'],boxprops = plotprops, capprops =plotprops,
                flierprops=flierprops,
                whiskerprops=plotprops)
    
    # stats    
    s01,p01 = ranksums(SnmN0h,SnmN12h)
    s12,p12 = ranksums(SnmN12h,SnmN24h)
    s02,p02 = ranksums(SnmN0h,SnmN24h)

    h01= np.max([np.max(SnmN0h),np.max(SnmN12h)])
    h12= np.max([np.max(SnmN24h),np.max(SnmN12h)])
    h02= np.max([np.max(SnmN0h),np.max(SnmN24h)])
    
    hmax = np.max([h01,h02,h12])
    step = hmax*0.15

    ax3.plot([1, 2], [hmax+step ,hmax+step], 'w-',zorder=0)
    ax3.text(1.5,hmax+step*1.3,'p = ' + str(round(p01*1000)/1000), ha='center',fontsize='small')
    
    ax3.plot([2, 3], [hmax+step*2 ,hmax+step*2], 'w-',zorder=0)
    ax3.text(2.5,hmax+step*2.3,'p = ' + str(round(p12*1000)/1000), ha='center',fontsize='small')
    
    ax3.plot([1, 3], [hmax+step*3 ,hmax+step*3], 'w-',zorder=0)
    ax3.text(2,hmax+step*3.3,'p = ' + str(round(p02*1000)/1000), ha='center',fontsize='small')
    
    ax3.set_ylim(top=hmax+step*4)
    
    fig3.savefig(P + '\\Figures\\SimiPlots\\Boxplot_DistancesToNormMean' + Title)

    if not Plots:
        plt.close(fig2)
        plt.close(fig3)

    if Plots:
        plt.show()
        
        
    
    
    return

print('Function defined.')
    

## Enter data and run

### Data details

Scale210903 = 1.94 # Spatial scale (µm/px) for 16X 03-09-2021
FPH210903 = 2 # Frames per hour

Scale210914 = 1.24 # Spatial scale (µm/px) for 25X 14-09-2021
FPH210914 = 2 # Frames per hour

Scale210927 = 1.24 # Spatial scale (µm/px) for 25X 27-09-2021
FPH210927 = 2 # Frames per hour

Scale211022 = 0.97 # Spatial scale (µm/px) for 32X 27-09-2021
FPH211022 = 2 # Frames per hour

Scale211105 = 1.55 # Spatial scale (µm/px) for 20X 05-11-2021
FPH211105 = 2 # Frames per hour

Scale211222 = 1.55 # Spatial scale (µm/px) for 20X 22-12-2021
FPH211222 = 2 # Frames per hour

Scale220107 = 1.94 # Spatial scale (µm/px) for 16X 07-01-2022
FPH220107 = 2 # Frames per hour

# Paths to data
P210903_1 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210903_V5_TestFlux\1mlh'
P210903_5 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210903_V5_TestFlux\500ulh'
P210914_Ct = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210914_V5_Manitol125mM\Ctrl'
P210914_M125 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210914_V5_Manitol125mM\Manitol125'
P210927_1 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210927_V5_TestFlux\1mlh'
P210927_5 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\210927_V5_TestFlux\500ulh'
P211022_Ct1 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211022_DV6_Ctrls\Ctrl1'
P211022_Ct2 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211022_DV6_Ctrls\Ctrl2'
P211105_Deg_Bsa = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211105_DV6_Degas_Ctrls\BSA'
P211105_Degas = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211105_DV6_Degas_Ctrls\NoCoat'
P211222_Deg_Bsa = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211222_DV6_Degas_Ctrls\BSA'
P211222_Degas = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\211222_DV6_Degas_Ctrls\NoCoat'
P220107_S3 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\220107_DV6_Stade1&3\Stade3'
P220107_S1 = r'D:\Users\laplaud\Desktop\PostDoc\Data\Microflu\220107_DV6_Stade1&3\Stade1'


# experiments of 03-09-2021, Puce V5, Ctrl with 1ml/heure medium flow
StackList210903_1 = ['PPG1','PPG2','PPG3','PPG6','PPG7','PPG10','PPG11','PPG12','PPG13','PPG14','PPG17','PPG19',
                   'PPG21','PPG22'] # 'PPG4','PPG5','PPG8','PPG9','PPG15','PPG16','PPG18','PPG20'

# experiments of 03-09-2021, Puce V5, Ctrl with 500ul/heure medium flow
StackList210903_5 = ['PPG3','PPG5','PPG6','PPG7','PPG8','PPG9','PPG10','PPG11'] # 'PPG1','PPG2','PPG4'

# experiments of 14-09-2021, Puce V5, Ctrl with 1ml/heure medium flow
StackList210914_Ct = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG9','PPG10','PPG11'] # 'PPG8','PPG12'

# experiments of 14-09-2021, Puce V5, 125mM manitol with 1ml/heure medium flow
StackList210914_M125 = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG7','PPG8','PPG9'] # 'PPG6'


# experiments of 27-09-2021, Puce V5, Ctrl with 1ml/heure medium flow
StackList210927_1 = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG7','PPG10',
                   'PPG11','PPG12','PPG13','PPG14','PPG15',] # 'PPG6','PPG8','PPG9'

# experiments of 27-09-2021, Puce V5, Ctrl with 500ul/heure medium flow
StackList210927_5 = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG8','PPG9','PPG10','PPG11','PPG13'] #'PPG7','PPG12'


# experiments of 22-10-2021, double puce V6, Ctrl 1 with 500ul/heure medium flow
StackList211022_Ct1 = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG8','PPG9',
                   'PPG11','PPG12','PPG13','PPG14','PPG15','PPG16','PPG17','PPG18','PPG19',
                       'PPG20','PPG21','PPG22','PPG23','PPG24','PPG25'] 
                    # 'PPG10', 
    
# experiments of 22-10-2021, double puce V6, Ctrl 2 with 500ul/heure medium flow
StackList211022_Ct2 = ['PPG1','PPG2','PPG4','PPG5','PPG6','PPG7','PPG8','PPG9','PPG10',
                   'PPG11','PPG12','PPG13','PPG15','PPG16','PPG18','PPG19','PPG20',
                   'PPG22','PPG23','PPG24','PPG26','PPG27','PPG28','PPG29']
                    # 'PPG3',  'PPG14', 'PPG17', 'PPG21', 'PPG25',


# experiments of 05-11-2021, double puce V6, Degased medium, BSA coated channel with 500ul/heure medium flow
StackList211105_Deg_Bsa = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG9','PPG10',
                     'PPG11', 'PPG13','PPG16','PPG18','PPG19','PPG20',
                   'PPG21','PPG23','PPG24','PPG25','PPG26','PPG27','PPG28','PPG29','PPG30']  #  
                       # PPG12, PPG8 dead , PPG15, PPG17 dead notch, PPG22 dead parts, PPG14 folded lost attach

# experiments of 05-11-2021, double puce V6, Degased medium  with 500ul/heure medium flow
StackList211105_Degas = ['PPG1','PPG2','PPG3','PPG4','PPG6','PPG7','PPG8','PPG9','PPG10',
                   'PPG11','PPG12','PPG13','PPG14','PPG16','PPG17','PPG18','PPG19','PPG20'] # 
                # PPG3 dead attach, PPG5 dead notch, 'PPG15' dead attach

    
# experiments of 22-12-2021, double puce V6, degased medium, bsa coating with 500ul/heure medium flow
# /!\ /!\ Start 30 minutes later than usual, first image removed
StackList211222_Deg_Bsa = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG8','PPG9','PPG10',
                   'PPG11','PPG12','PPG14','PPG15','PPG16','PPG17','PPG18','PPG19','PPG20',
                   'PPG21','PPG22','PPG23','PPG24','PPG25','PPG26','PPG28','PPG29','PPG30',
                          'PPG31'] # PPG22 ,'PPG13' touche le bord de l'image # ,'PPG27' bad contouring


# experiments of 22-12-2021, double puce V6, degased medium, no coating with 500ul/heure medium flow
# /!\ /!\ Start 30 minutes later than usual, first image removed
StackList211222_Degas = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG9','PPG10',
                   'PPG11','PPG12','PPG13','PPG14','PPG15','PPG16','PPG17','PPG18','PPG19','PPG20',
                   'PPG21','PPG22','PPG23','PPG24','PPG25','PPG26','PPG27','PPG28','PPG29','PPG30']
                    # PPG8 trop plié


# experiments of 07-01-2022, double puce V6, degased medium, no coating with 500ul/heure medium flow
# Stade1 (close to the center) cups
StackList220107_S1 = ['PPG1','PPG2','PPG3','PPG5','PPG6','PPG7','PPG9','PPG10',
                   'PPG11','PPG12','PPG13','PPG14','PPG15','PPG16','PPG17','PPG18','PPG19','PPG20',
                   'PPG21','PPG22','PPG23','PPG24','PPG25','PPG26','PPG27']


# experiments of 07-01-2022, double puce V6, degased medium, no coating with 500ul/heure medium flow
# Stade3 (close to the edge) cups
StackList220107_S3 = ['PPG1','PPG2','PPG3','PPG4','PPG5','PPG6','PPG7','PPG8','PPG9','PPG10',
                   'PPG11','PPG12','PPG13','PPG14','PPG15','PPG16','PPG17','PPG18','PPG19','PPG20']


print('Data choice made.')


### Loading data 

## Expe 210903 500ulh
ContourData210903_5 = pd.read_csv(P210903_5 + '\\ContourData210903_500ulh_Param.csv', index_col = 'Ind')
GlobalData210903_5 = pd.read_csv(P210903_5 + '\\GlobalData210903_500ulh_Aligned.csv', index_col = 'Ind') 

## Expe 210903 1mlh
ContourData210903_1 = pd.read_csv(P210903_1 + '\\ContourData210903_1mlh_Param.csv', index_col = 'Ind')
GlobalData210903_1 = pd.read_csv(P210903_1 + '\\GlobalData210903_1mlh_Aligned.csv', index_col = 'Ind')  

## Expe 210914 Ctrl
ContourData210914_Ct = pd.read_csv(P210914_Ct + '\\ContourData210914_Ct_Param.csv', index_col = 'Ind')
GlobalData210914_Ct = pd.read_csv(P210914_Ct + '\\GlobalData210914_Ct_Aligned.csv', index_col = 'Ind') 

## Expe 210914 manitol 125mM
ContourData210914_M125 = pd.read_csv(P210914_M125 + '\\ContourData210914_M125_Param.csv', index_col = 'Ind')
GlobalData210914_M125 = pd.read_csv(P210914_M125 + '\\GlobalData210914_M125_Aligned.csv', index_col = 'Ind')  

## Expe 210927 500ulh
ContourData210927_5 = pd.read_csv(P210927_5 + '\\ContourData210927_500ulh_Param.csv', index_col = 'Ind')
GlobalData210927_5 = pd.read_csv(P210927_5 + '\\GlobalData210927_500ulh_Aligned.csv', index_col = 'Ind') 

## Expe 210927 1mlh
ContourData210927_1 = pd.read_csv(P210927_1 + '\\ContourData210927_1mlh_Param.csv', index_col = 'Ind')
GlobalData210927_1 = pd.read_csv(P210927_1 + '\\GlobalData210927_1mlh_Aligned.csv', index_col = 'Ind')  

## Expe 211022 Ct1
ContourData211022_Ct1 = pd.read_csv(P211022_Ct1 + '\\ContourData211022_Ct1_Param.csv', index_col = 'Ind')
GlobalData211022_Ct1 = pd.read_csv(P211022_Ct1 + '\\GlobalData211022_Ct1_Aligned.csv', index_col = 'Ind')  

## Expe 211022 Ct2
ContourData211022_Ct2 = pd.read_csv(P211022_Ct2 + '\\ContourData211022_Ct2_Param.csv', index_col = 'Ind')
GlobalData211022_Ct2 = pd.read_csv(P211022_Ct2 + '\\GlobalData211022_Ct2_Aligned.csv', index_col = 'Ind')   

## Expe 211105 Degas + BSA
ContourData211105_Deg_Bsa = pd.read_csv(P211105_Deg_Bsa + '\\ContourData211105_Deg_Bsa_Param.csv', index_col = 'Ind')
GlobalData211105_Deg_Bsa = pd.read_csv(P211105_Deg_Bsa + '\\GlobalData211105_Deg_Bsa_Aligned.csv', index_col = 'Ind')  

## Expe 211105 Degas 
ContourData211105_Degas = pd.read_csv(P211105_Degas + '\\ContourData211105_Degas_Param.csv', index_col = 'Ind')
GlobalData211105_Degas = pd.read_csv(P211105_Degas + '\\GlobalData211105_Degas_Aligned.csv', index_col = 'Ind')  

print('Parametrised contours loaded. \n')


### Adding manual info to global data

def AddExcelSheet(GD, StackList, SheetPath):
    Sheet = pd.read_excel(SheetPath,index_col=0)
    for c in Sheet.columns:
        for s in StackList:
            GD.loc[(GD.index == s) & (GD['Img'].values == 0), c] = bool(Sheet.loc[s,c])

    return GD

## Expe 211022 Ct1
GlobalData211022_Ct1 = AddExcelSheet(GlobalData211022_Ct1, StackList211022_Ct1, P211022_Ct1+'\PPGInfo.xlsx')
## Expe 211022 Ct2
GlobalData211022_Ct2 = AddExcelSheet(GlobalData211022_Ct2, StackList211022_Ct2, P211022_Ct2+'\PPGInfo.xlsx')

### Compute contour characteristics

#### Mean contours

## Expe 210903
print('\033[1m' + '\033[4m' + '\nAnalyzing 500ulh experiments from 2021-09-03 :\n' + '\033[0m')
ContourData210903_5 = computeMeanContourTime(StackList210903_5,ContourData210903_5,GlobalData210903_5)

print('\033[1m' + '\033[4m' + '\nAnalyzing 1mlh experiments from 2021-09-03 :\n' + '\033[0m')
ContourData210903_1 = computeMeanContourTime(StackList210903_1,ContourData210903_1,GlobalData210903_1)

# Expe 210914
print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl experiments from 2021-09-14 :\n' + '\033[0m')
ContourData210914_Ct = computeMeanContourTime(StackList210914_Ct,ContourData210914_Ct,GlobalData210914_Ct)

print('\033[1m' + '\033[4m' + '\nAnalyzing manitol 125mM experiments from 2021-09-14 :\n' + '\033[0m')
ContourData210914_M125 = computeMeanContourTime(StackList210914_M125,ContourData210914_M125,GlobalData210914_M125)

## Expe 210927
print('\033[1m' + '\033[4m' + '\nAnalyzing 500ulh experiments from 2021-09-27 :\n' + '\033[0m')
ContourData210927_5 = computeMeanContourTime(StackList210927_5,ContourData210927_5,GlobalData210927_5)

print('\033[1m' + '\033[4m' + '\nAnalyzing 1mlh experiments from 2021-09-27 :\n' + '\033[0m')
ContourData210927_1 = computeMeanContourTime(StackList210927_1,ContourData210927_1,GlobalData210927_1)

print('\033[1m' + '\033[4m' + '\nAnalyzing 1mlh experiments from 2021-09-27 :\n' + '\033[0m')
ContourData210927_1 = computeMeanContourTime(StackList210927_1,ContourData210927_1,GlobalData210927_1)

print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl1 experiments from 2021-10-22 :\n' + '\033[0m')
ContourData211022_Ct1 = computeMeanContourTime(StackList211022_Ct1,ContourData211022_Ct1,GlobalData211022_Ct1)

print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl2 experiments from 2021-10-22 :\n' + '\033[0m')
ContourData211022_Ct2 = computeMeanContourTime(StackList211022_Ct2,ContourData211022_Ct2,GlobalData211022_Ct2)

print('\033[1m' + '\033[4m' + '\nAnalyzing Degas + BSA experiments from 2021-11-05 :\n' + '\033[0m')
ContourData211105_Deg_Bsa = computeMeanContourTime(StackList211105_Deg_Bsa,ContourData211105_Deg_Bsa,GlobalData211105_Deg_Bsa)

print('\033[1m' + '\033[4m' + '\nAnalyzing Degas experiments from 2021-11-05 :\n' + '\033[0m')
ContourData211105_Degas = computeMeanContourTime(StackList211105_Degas,ContourData211105_Degas,GlobalData211105_Degas)

print('Mean contours done. \n')

#### Compute similarities and growth

## Expe 210903
print('\033[1m' + '\033[4m' + '\nAnalyzing 500ulh experiments from 2021-09-03 :\n' + '\033[0m')
ContourData210903_5,GrowthMat210903_5,SimMat210903_5 = computeGrowthAndSimi(StackList210903_5,ContourData210903_5,GlobalData210903_5)

print('\033[1m' + '\033[4m' + '\nAnalyzing 1mlh experiments from 2021-09-03 :\n' + '\033[0m')
ContourData210903_1,GrowthMat210903_1,SimMat210903_1 = computeGrowthAndSimi(StackList210903_1,ContourData210903_1,GlobalData210903_1)

# expe 210914
print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl experiments from 2021-09-14 :\n' + '\033[0m')
ContourData210914_Ct,GrowthMat210914_Ct,SimMat210914_Ct = computeGrowthAndSimi(StackList210914_Ct,ContourData210914_Ct,GlobalData210914_Ct)

print('\033[1m' + '\033[4m' + '\nAnalyzing manitol 125mM experiments from 2021-09-14 :\n' + '\033[0m')
ContourData210914_M125,GrowthMat210914_M125,SimMat210914_M125 = computeGrowthAndSimi(StackList210914_M125,ContourData210914_M125,GlobalData210914_M125)

## Expe 210927
print('\033[1m' + '\033[4m' + '\nAnalyzing 500ulh experiments from 2021-09-27 :\n' + '\033[0m')
ContourData210927_5,GrowthMat210927_5,SimMat210927_5 = computeGrowthAndSimi(StackList210927_5,ContourData210927_5,GlobalData210927_5)

print('\033[1m' + '\033[4m' + '\nAnalyzing 1mlh experiments from 2021-09-27 :\n' + '\033[0m')
ContourData210927_1,GrowthMat210927_1,SimMat210927_1 = computeGrowthAndSimi(StackList210927_1,ContourData210927_1,GlobalData210927_1)


## Expe 211022
print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl1 experiments from 2021-10-22 :\n' + '\033[0m')
ContourData211022_Ct1,GrowthMat211022_Ct1,SimMat211022_Ct1 = computeGrowthAndSimi(StackList211022_Ct1,ContourData211022_Ct1,GlobalData211022_Ct1)
## Expe 211022
print('\033[1m' + '\033[4m' + '\nAnalyzing Ctrl2 experiments from 2021-10-22 :\n' + '\033[0m')
ContourData211022_Ct2,GrowthMat211022_Ct2,SimMat211022_Ct2 = computeGrowthAndSimi(StackList211022_Ct2,ContourData211022_Ct2,GlobalData211022_Ct2)

## Expe 211105
print('\033[1m' + '\033[4m' + '\nAnalyzing Degas + BSA experiments from 2021-11-05 :\n' + '\033[0m')
ContourData211105_Deg_Bsa,GrowthMat211105_Deg_Bsa,SimMat211105_Deg_Bsa = computeGrowthAndSimi(StackList211105_Deg_Bsa,ContourData211105_Deg_Bsa,GlobalData211105_Deg_Bsa)

print('\033[1m' + '\033[4m' + '\nAnalyzing Degas experiments from 2021-11-05 :\n' + '\033[0m')
ContourData211105_Degas,GrowthMat211105_Degas,SimMat211105_Degas = computeGrowthAndSimi(StackList211105_Degas,ContourData211105_Degas,GlobalData211105_Degas)


print('Similarities and growth done. \n')

###  Save quantifications

## Saving
SaveList210903_5 = StackList210903_5 + ['meanContTime', 'normMeanContTime']# , 'meanContPDF', 'meanContArea'
GlobalData210903_5 = GlobalData210903_5.loc[StackList210903_5]
ContourData210903_5 = ContourData210903_5.loc[SaveList210903_5]
GlobalData210903_5.to_csv(P210903_5 + '\\GlobalData210903_500ulh_Quantif.csv',index_label = 'Ind')
ContourData210903_5.to_csv(P210903_5 + '\\ContourData210903_500ulh_Quantif.csv',index_label = 'Ind')
np.save(P210903_5 + '\\GrowthMat210903_500ulh', GrowthMat210903_5)
np.save(P210903_5 + '\\SimMat210903_500ulh', SimMat210903_5)

SaveList210903_1 = StackList210903_1 + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData210903_1 = GlobalData210903_1.loc[StackList210903_1]
ContourData210903_1 = ContourData210903_1.loc[SaveList210903_1]
GlobalData210903_1.to_csv(P210903_1 + '\\GlobalData210903_1mlh_Quantif.csv',index_label = 'Ind')
ContourData210903_1.to_csv(P210903_1 + '\\ContourData210903_1mlh_Quantif.csv',index_label = 'Ind')
np.save(P210903_1 + '\\GrowthMat210903_1mlh', GrowthMat210903_1)
np.save(P210903_1 + '\\SimMat210903_1mlh', SimMat210903_1)

SaveList210914_M125 = StackList210914_M125 + ['meanContTime', 'normMeanContTime']# , 'meanContPDF', 'meanContArea'
GlobalData210914_M125 = GlobalData210914_M125.loc[StackList210914_M125]
ContourData210914_M125 = ContourData210914_M125.loc[SaveList210914_M125]
GlobalData210914_M125.to_csv(P210914_M125 + '\\GlobalData210914_M125_Quantif.csv',index_label = 'Ind')
ContourData210914_M125.to_csv(P210914_M125 + '\\ContourData210914_M125_Quantif.csv',index_label = 'Ind')
np.save(P210914_M125 + '\\GrowthMat210914_M125', GrowthMat210914_M125)
np.save(P210914_M125 + '\\SimMat210914_M125', SimMat210914_M125)

SaveList210914_Ct = StackList210914_Ct +  ['meanContTime', 'normMeanContTime']# , 'meanContPDF', 'meanContArea'
GlobalData210914_Ct = GlobalData210914_Ct.loc[StackList210914_Ct]
ContourData210914_Ct = ContourData210914_Ct.loc[SaveList210914_Ct]
GlobalData210914_Ct.to_csv(P210914_Ct + '\\GlobalData210914_Ct_Quantif.csv',index_label = 'Ind')
ContourData210914_Ct.to_csv(P210914_Ct + '\\ContourData210914_Ct_Quantif.csv',index_label = 'Ind')
np.save(P210914_Ct + '\\GrowthMat210914_Ct', GrowthMat210914_Ct)
np.save(P210914_Ct + '\\SimMat210914_Ct', SimMat210914_Ct)

SaveList210927_5 = StackList210927_5 + ['meanContTime', 'normMeanContTime']# , 'meanContPDF', 'meanContArea'
GlobalData210927_5 = GlobalData210927_5.loc[StackList210927_5]
ContourData210927_5 = ContourData210927_5.loc[SaveList210927_5]
GlobalData210927_5.to_csv(P210927_5 + '\\GlobalData210927_500ulh_Quantif.csv',index_label = 'Ind')
ContourData210927_5.to_csv(P210927_5 + '\\ContourData210927_500ulh_Quantif.csv',index_label = 'Ind')
np.save(P210927_5 + '\\GrowthMat210927_500ulh', GrowthMat210927_5)
np.save(P210927_5 + '\\SimMat210927_500ulh', SimMat210927_5)

SaveList210927_1 = StackList210927_1 + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData210927_1 = GlobalData210927_1.loc[StackList210927_1]
ContourData210927_1 = ContourData210927_1.loc[SaveList210927_1]
GlobalData210927_1.to_csv(P210927_1 + '\\GlobalData210927_1mlh_Quantif.csv',index_label = 'Ind')
ContourData210927_1.to_csv(P210927_1 + '\\ContourData210927_1mlh_Quantif.csv',index_label = 'Ind')
np.save(P210927_1 + '\\GrowthMat210927_1mlh', GrowthMat210927_1)
np.save(P210927_1 + '\\SimMat210927_1mlh', SimMat210927_1)

SaveList211022_Ct1 = StackList211022_Ct1 + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData211022_Ct1 = GlobalData211022_Ct1.loc[StackList211022_Ct1]
ContourData211022_Ct1 = ContourData211022_Ct1.loc[SaveList211022_Ct1]
GlobalData211022_Ct1.to_csv(P211022_Ct1 + '\\GlobalData211022_Ct1_Quantif.csv',index_label = 'Ind')
ContourData211022_Ct1.to_csv(P211022_Ct1 + '\\ContourData211022_Ct1_Quantif.csv',index_label = 'Ind')
np.save(P211022_Ct1 + '\\GrowthMat211022_Ct1', GrowthMat211022_Ct1)
np.save(P211022_Ct1 + '\\SimMat211022_Ct1', SimMat211022_Ct1)

SaveList211022_Ct2 = StackList211022_Ct2 + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData211022_Ct2 = GlobalData211022_Ct2.loc[StackList211022_Ct2]
ContourData211022_Ct2 = ContourData211022_Ct2.loc[SaveList211022_Ct2]
GlobalData211022_Ct2.to_csv(P211022_Ct2 + '\\GlobalData211022_Ct2_Quantif.csv',index_label = 'Ind')
ContourData211022_Ct2.to_csv(P211022_Ct2 + '\\ContourData211022_Ct2_Quantif.csv',index_label = 'Ind')
np.save(P211022_Ct2 + '\\GrowthMat211022_Ct2', GrowthMat211022_Ct2)
np.save(P211022_Ct2 + '\\SimMat211022_Ct2', SimMat211022_Ct2)

SaveList211105_Deg_Bsa = StackList211105_Deg_Bsa + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData211105_Deg_Bsa = GlobalData211105_Deg_Bsa.loc[StackList211105_Deg_Bsa]
ContourData211105_Deg_Bsa = ContourData211105_Deg_Bsa.loc[SaveList211105_Deg_Bsa]
GlobalData211105_Deg_Bsa.to_csv(P211105_Deg_Bsa + '\\GlobalData211105_Deg_Bsa_Quantif.csv',index_label = 'Ind')
ContourData211105_Deg_Bsa.to_csv(P211105_Deg_Bsa + '\\ContourData211105_Deg_Bsa_Quantif.csv',index_label = 'Ind')
np.save(P211105_Deg_Bsa + '\\GrowthMat211105_Deg_Bsa', GrowthMat211105_Deg_Bsa)
np.save(P211105_Deg_Bsa + '\\SimMat211105_Deg_Bsa', SimMat211105_Deg_Bsa)

SaveList211105_Degas = StackList211105_Degas + ['meanContTime', 'normMeanContTime'] # , 'meanContPDF', 'meanContArea'
GlobalData211105_Degas = GlobalData211105_Degas.loc[StackList211105_Degas]
ContourData211105_Degas = ContourData211105_Degas.loc[SaveList211105_Degas]
GlobalData211105_Degas.to_csv(P211105_Degas + '\\GlobalData211105_Degas_Quantif.csv',index_label = 'Ind')
ContourData211105_Degas.to_csv(P211105_Degas + '\\ContourData211105_Degas_Quantif.csv',index_label = 'Ind')
np.save(P211105_Degas + '\\GrowthMat211105_Degas', GrowthMat211105_Degas)
np.save(P211105_Degas + '\\SimMat211105_Degas', SimMat211105_Degas)

print('Quantifications saved. \n')

### Plot results

#### Load quantifications

## Expe 210903
ContourData210903_5 = pd.read_csv(P210903_5 + '\\ContourData210903_500ulh_Quantif.csv', index_col = 'Ind')
GlobalData210903_5 = pd.read_csv(P210903_5 + '\\GlobalData210903_500ulh_Quantif.csv', index_col = 'Ind')    
GrowthMat210903_5 = np.load(P210903_5 + '\\GrowthMat210903_500ulh.npy')    
SimMat210903_5 = np.load(P210903_5 + '\\SimMat210903_500ulh.npy')

ContourData210903_1 = pd.read_csv(P210903_1 + '\\ContourData210903_1mlh_Quantif.csv', index_col = 'Ind')
GlobalData210903_1 = pd.read_csv(P210903_1 + '\\GlobalData210903_1mlh_Quantif.csv', index_col = 'Ind')    
GrowthMat210903_1 = np.load(P210903_1 + '\\GrowthMat210903_1mlh.npy') 
SimMat210903_1 = np.load(P210903_1 + '\\SimMat210903_1mlh.npy')

## Expe 210914 ctrl
ContourData210914_Ct = pd.read_csv(P210914_Ct + '\\ContourData210914_Ct_Quantif.csv', index_col = 'Ind')
GlobalData210914_Ct = pd.read_csv(P210914_Ct + '\\GlobalData210914_Ct_Quantif.csv', index_col = 'Ind')    
GrowthMat210914_Ct = np.load(P210914_Ct + '\\GrowthMat210914_Ct.npy')   
SimMat210914_Ct = np.load(P210914_Ct + '\\SimMat210914_Ct.npy')

## Expe 210914 manitol 125mM
ContourData210914_M125 = pd.read_csv(P210914_M125 + '\\ContourData210914_M125_Quantif.csv', index_col = 'Ind')
GlobalData210914_M125 = pd.read_csv(P210914_M125 + '\\GlobalData210914_M125_Quantif.csv', index_col = 'Ind')    
GrowthMat210914_M125 = np.load(P210914_M125 + '\\GrowthMat210914_M125.npy')   
SimMat210914_M125 = np.load(P210914_M125 + '\\SimMat210914_M125.npy')

## Expe 210927
ContourData210927_5 = pd.read_csv(P210927_5 + '\\ContourData210927_500ulh_Quantif.csv', index_col = 'Ind')
GlobalData210927_5 = pd.read_csv(P210927_5 + '\\GlobalData210927_500ulh_Quantif.csv', index_col = 'Ind')    
GrowthMat210927_5 = np.load(P210927_5 + '\\GrowthMat210927_500ulh.npy')  
SimMat210927_5 = np.load(P210927_5 + '\\SimMat210927_500ulh.npy')

ContourData210927_1 = pd.read_csv(P210927_1 + '\\ContourData210927_1mlh_Quantif.csv', index_col = 'Ind')
GlobalData210927_1 = pd.read_csv(P210927_1 + '\\GlobalData210927_1mlh_Quantif.csv', index_col = 'Ind')    
GrowthMat210927_1 = np.load(P210927_1 + '\\GrowthMat210927_1mlh.npy')    
SimMat210927_1 = np.load(P210927_1 + '\\SimMat210927_1mlh.npy')

## Expe 211022
ContourData211022_Ct1 = pd.read_csv(P211022_Ct1 + '\\ContourData211022_Ct1_Quantif.csv', index_col = 'Ind')
GlobalData211022_Ct1 = pd.read_csv(P211022_Ct1 + '\\GlobalData211022_Ct1_Quantif.csv', index_col = 'Ind')    
GrowthMat211022_Ct1 = np.load(P211022_Ct1 + '\\GrowthMat211022_Ct1.npy')  
SimMat211022_Ct1 = np.load(P211022_Ct1 + '\\SimMat211022_Ct1.npy')

ContourData211022_Ct2 = pd.read_csv(P211022_Ct2 + '\\ContourData211022_Ct2_Quantif.csv', index_col = 'Ind')
GlobalData211022_Ct2 = pd.read_csv(P211022_Ct2 + '\\GlobalData211022_Ct2_Quantif.csv', index_col = 'Ind')    
GrowthMat211022_Ct2 = np.load(P211022_Ct2 + '\\GrowthMat211022_Ct2.npy')  
SimMat211022_Ct2 = np.load(P211022_Ct2 + '\\SimMat211022_Ct2.npy')

ContourData211105_Deg_Bsa = pd.read_csv(P211105_Deg_Bsa + '\\ContourData211105_Deg_Bsa_Quantif.csv', index_col = 'Ind')
GlobalData211105_Deg_Bsa = pd.read_csv(P211105_Deg_Bsa + '\\GlobalData211105_Deg_Bsa_Quantif.csv', index_col = 'Ind')    
GrowthMat211105_Deg_Bsa = np.load(P211105_Deg_Bsa + '\\GrowthMat211105_Deg_Bsa.npy')  
SimMat211105_Deg_Bsa = np.load(P211105_Deg_Bsa + '\\SimMat211105_Deg_Bsa.npy')

ContourData211105_Degas = pd.read_csv(P211105_Degas + '\\ContourData211105_Degas_Quantif.csv', index_col = 'Ind')
GlobalData211105_Degas = pd.read_csv(P211105_Degas + '\\GlobalData211105_Degas_Quantif.csv', index_col = 'Ind')    
GrowthMat211105_Degas = np.load(P211105_Degas + '\\GrowthMat211105_Degas.npy')  
SimMat211105_Degas = np.load(P211105_Degas + '\\SimMat211105_Degas.npy')

print('Loading OK.')

#### Plots

""" plotArea(StackList, GD, P, CCs, Title)
    plotShapes(P, StackList,CD,GD, figname)
    plotGrowth(P,StackList,CD,GrowthMat)
    plotSimi(P,StackList,CD,SimMatn Title) """

plotShapes(P210903_5,StackList210903_5,ContourData210903_5,GlobalData210903_5,'Expe_210903_500ulh', showplots = False)
plotGrowth(P210903_5,['meanContTime'] + StackList210903_5,ContourData210903_5,GrowthMat210903_5, showplots = False)
plotSimi(P210903_5,StackList210903_5,ContourData210903_5,SimMat210903_5,'Expe_210903_500ulh', showplots = False)

plotShapes(P210903_1,StackList210903_1,ContourData210903_1,GlobalData210903_1,'Expe_210903_1mlh', showplots = False)
plotGrowth(P210903_1,['meanContTime'] + StackList210903_1,ContourData210903_1,GrowthMat210903_1, showplots = False)
plotSimi(P210903_1,StackList210903_1,ContourData210903_1,SimMat210903_1,'Expe_210903_1mlh', showplots = False)

plotShapes(P210914_Ct,StackList210914_Ct,ContourData210914_Ct,GlobalData210914_Ct,'Expe_210914_Ct', showplots = False)
plotGrowth(P210914_Ct,['meanContTime'] + StackList210914_Ct,ContourData210914_Ct,GrowthMat210914_Ct, showplots = False)
plotSimi(P210914_Ct,StackList210914_Ct,ContourData210914_Ct,SimMat210914_Ct,'Expe_210914_Ct', showplots = False)

plotShapes(P210914_M125,StackList210914_M125,ContourData210914_M125,GlobalData210914_M125,'Expe_210914_M125', showplots = False)
plotGrowth(P210914_M125,['meanContTime'] + StackList210914_M125,ContourData210914_M125,GrowthMat210914_M125, showplots = False)
plotSimi(P210914_M125,StackList210914_M125,ContourData210914_M125,SimMat210914_M125,'Expe_210914_M125', showplots = False)

plotShapes(P210927_5,StackList210927_5,ContourData210927_5,GlobalData210927_5,'Expe_210927_500ulh', showplots = False)
plotGrowth(P210927_5,['meanContTime'] + StackList210927_5,ContourData210927_5,GrowthMat210927_5, showplots = False)
plotSimi(P210927_5,StackList210927_5,ContourData210927_5,SimMat210927_5, 'Expe_210927_500ulh', showplots = False)

plotShapes(P210927_1,StackList210927_1,ContourData210927_1,GlobalData210927_1,'Expe_210927_1mlh', showplots = False)
plotGrowth(P210927_1,['meanContTime'] + StackList210927_1,ContourData210927_1,GrowthMat210927_1, showplots = False)
plotSimi(P210927_1,StackList210927_1,ContourData210927_1,SimMat210927_1,'Expe_210927_1mlh', showplots = False)

plotShapes(P211022_Ct1,StackList211022_Ct1,ContourData211022_Ct1,GlobalData211022_Ct1,'Expe_211022_Ct1', showplots = True)
plotGrowth(P211022_Ct1,['meanContTime'] + StackList211022_Ct1,ContourData211022_Ct1,GrowthMat211022_Ct1, showplots = False)
plotSimi(P211022_Ct1,StackList211022_Ct1,ContourData211022_Ct1,SimMat211022_Ct1,'Expe_211022_Ct1', showplots = False)

plotShapes(P211022_Ct2,StackList211022_Ct2,ContourData211022_Ct2,GlobalData211022_Ct2,'Expe_211022_Ct2', showplots = True)
plotGrowth(P211022_Ct2,['meanContTime'] + StackList211022_Ct2,ContourData211022_Ct2,GrowthMat211022_Ct2, showplots = False)
plotSimi(P211022_Ct2,StackList211022_Ct2,ContourData211022_Ct2,SimMat211022_Ct2,'Expe_211022_Ct2', showplots = False)
         
plotShapes(P211105_Deg_Bsa,StackList211105_Deg_Bsa,ContourData211105_Deg_Bsa,GlobalData211105_Deg_Bsa,'Expe_211105_Deg_Bsa', showplots = True)
plotGrowth(P211105_Deg_Bsa,['meanContTime'] + StackList211105_Deg_Bsa,ContourData211105_Deg_Bsa,GrowthMat211105_Deg_Bsa, showplots = False)
plotSimi(P211105_Deg_Bsa,StackList211105_Deg_Bsa,ContourData211105_Deg_Bsa,SimMat211105_Deg_Bsa,'Expe_211105_Deg_Bsa', showplots = False)

plotShapes(P211105_Degas,StackList211105_Degas,ContourData211105_Degas,GlobalData211105_Degas,'Expe_211105_Degas', showplots = True)
plotGrowth(P211105_Degas,['meanContTime'] + StackList211105_Degas,ContourData211105_Degas,GrowthMat211105_Degas, showplots = False)
plotSimi(P211105_Degas,StackList211105_Degas,ContourData211105_Degas,SimMat211105_Degas,'Expe_211105_Degas', showplots = False)

## Test Zone
